requirements.txt
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
pypdf
chromadb
streamlit
python-dotenv
google-search-results
pyngrok
